+++ waves/authentication/admin.py
@@ -2,13 +2,13 @@ from __future__ import unicode_literals
 
 from django.contrib import admin
 
-from waves.authentication.models import ApiKey
+from waves.authentication.models import WavesApiUser
 
 
 class ApiKeyAdmin(admin.ModelAdmin):
-    list_display = ('key', 'user', 'created')
+    list_display = ('user', 'key', 'created')
     fields = ('user',)
     ordering = ('-created',)
 
 
-admin.site.register(ApiKey, ApiKeyAdmin)
+admin.site.register(WavesApiUser, ApiKeyAdmin)