+++ CHANGES.md
@@ -4,7 +4,8 @@ CHANGELOG
 Version 1.1.9.x - 2018-03-10
 ----------------------------
 
-- [SITES] Added multi-site association allow returning site url prefix for jobs urls
+- [USERS] - Added API USER class, unifying authentication, url redirection dedicated to REST Api users 
+    - Added multi-site association allow returning site url prefix for jobs urls
     - Test if job.client user has a 'site' property, in such case, retrieve domain name to generate Job / JobOutputs link
 - [LOGGING] JobLogging: degraded mode, output to 'waves.errors' logger